Added readln & now with writeln they use the symtable.
There's definitely some border case thing going on, since readln is
the only procedure/function that takes something differetn than an
integer.

Note:
* maybe instead of obtaining the identifier name we can re-cast the
result of the expression to a pointer type? Although that would
introduce pointer arithmetic and it'd be annoying.
* it's really annoying that readln() in pascal is an exception in the
sense that any variable is passed as a pointer to it.